requirements.txt
playwright>=1.49.0
google-genai>=1.0.0
python-dotenv>=1.0.0
customtkinter>=5.2.0
easyocr>=1.7.0
